Add CLOSED_LOST deadline handler logic
Add CLOSED_LOST deadline handler logic. For this, schedule a deadline
when the Opportunity is opened in the Account aggregate. Furthermore,
add the endDate to the opportunity open(ed) command and event. Lastly,
handle the deadline in the Opportunity, publishing the ClosedLost event
as a result, which in turn adjusts the opportunity's stage. Closing: add
 some todos to clarify with Emmett

#init